+++ src/Products/MeetingCommunes/profiles/zcity/import_data.py
@@ -21,7 +21,7 @@ categories = []
 # no user
 groups = [OrgDescriptor('dirgen', 'Directeur Général', u'DG'),
           OrgDescriptor('secretariat', 'Secrétariat Général', u'Secr'),
-          OrgDescriptor('dirfin', 'Directeur Financier', u'DF'), ]
+          OrgDescriptor('dirfin', 'Directeur Financier', u'DF')]
 
 # Meeting configurations -------------------------------------------------------
 # College
@@ -50,9 +50,10 @@ collegeMeeting.usedMeetingAttributes = ['start_date',
 collegeMeeting.insertingMethodsOnAddItem = (
     {'insertingMethod': 'on_list_type', 'reverse': '0'},
     {'insertingMethod': 'on_proposing_groups', 'reverse': '0'})
-collegeMeeting.itemReferenceFormat = "python: 'COL/' + (here.hasMeeting() and " \
-                                     "here.restrictedTraverse('@@pm_unrestricted_methods').getLinkedMeetingDate().strftime('%Y%m%d') or '') " \
-                                     "+ '-' + str(here.getItemNumber(relativeTo='meeting', for_display=True))"
+collegeMeeting.itemReferenceFormat = \
+    "python: 'COL/' + (here.hasMeeting() and " \
+    "here.restrictedTraverse('@@pm_unrestricted_methods').getLinkedMeetingDate().strftime('%Y%m%d') or '') " \
+    "+ '-' + str(here.getItemNumber(relativeTo='meeting', for_display=True))"
 collegeMeeting.contentsKeptOnSentToOtherMC = ['annexes', 'decision_annexes', 'advices']
 collegeMeeting.itemWFValidationLevels = (
     {'state': 'itemcreated',
@@ -102,8 +103,8 @@ collegeMeeting.transitionsToConfirm = (
     'MeetingItem.backTo_presented_from_returned_to_proposing_group', 'MeetingItem.delay',
     'MeetingItem.backToValidated', 'MeetingItem.return_to_proposing_group')
 collegeMeeting.enabledAnnexesBatchActions = ['delete', 'download-annexes']
-collegeMeeting.dashboardItemsListingsFilters = ('c4', 'c6', 'c7', 'c8', 'c9', 'c10',
-                                              'c11', 'c13', 'c14', 'c15', 'c16', 'c29', 'c32')
+collegeMeeting.dashboardItemsListingsFilters = (
+    'c4', 'c6', 'c7', 'c8', 'c9', 'c10', 'c11', 'c13', 'c14', 'c15', 'c16', 'c29', 'c32')
 collegeMeeting.dashboardMeetingAvailableItemsFilters = ('c4', 'c11', 'c16', 'c29', 'c32')
 collegeMeeting.dashboardMeetingLinkedItemsFilters = ('c4', 'c6', 'c7', 'c11', 'c16', 'c19', 'c29', 'c32')
 collegeMeeting.selectableAdvisers = []
@@ -119,8 +120,9 @@ collegeMeeting.itemAdviceViewStates = ('proposed',
                                        'accepted_but_modified',
                                        'refused',
                                        'delayed')
-collegeMeeting.usedAdviceTypes = ['positive', 'positive_with_comments', 'positive_with_remarks', 'cautious',
-                                  'negative', 'negative_with_remarks', 'back_to_proposing_group', 'nil', 'read']
+collegeMeeting.usedAdviceTypes = [
+    'positive', 'positive_with_comments', 'positive_with_remarks', 'cautious',
+    'negative', 'negative_with_remarks', 'back_to_proposing_group', 'nil', 'read']
 collegeMeeting.keepAccessToItemWhenAdviceIsGiven = True
 collegeMeeting.transitionsReinitializingDelays = ['backToItemCreated']
 # use template file from profile examples_fr
@@ -135,7 +137,8 @@ agendaTemplate.pod_portal_types = ['Meeting']
 agendaTemplate.tal_condition = u'python:tool.isManager(cfg)'
 
 agendaTemplatePDF = PodTemplateDescriptor('oj-2', 'Ordre du jour')
-agendaTemplatePDF.pod_template_to_use = {'cfg_id': collegeMeeting.id, 'template_id': agendaTemplate.id}
+agendaTemplatePDF.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': agendaTemplate.id}
 agendaTemplatePDF.pod_formats = ['pdf']
 agendaTemplatePDF.pod_portal_types = ['Meeting']
 
@@ -147,39 +150,42 @@ decisionsTemplate.pod_portal_types = ['Meeting']
 decisionsTemplate.tal_condition = u'python:tool.isManager(cfg)'
 
 decisionsTemplatePDF = PodTemplateDescriptor('pv-2', 'Procès-verbal')
-decisionsTemplatePDF.pod_template_to_use = {'cfg_id': collegeMeeting.id, 'template_id': decisionsTemplate.id}
+decisionsTemplatePDF.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': decisionsTemplate.id}
 decisionsTemplatePDF.pod_formats = ['pdf']
 decisionsTemplatePDF.pod_portal_types = ['Meeting']
 
 itemTemplate = PodTemplateDescriptor('deliberation', 'Délibération')
 itemTemplate.is_reusable = True
 itemTemplate.odt_file = '../../examples_fr/templates/deliberation.odt'
-itemTemplate.pod_formats = ['docx' ]
+itemTemplate.pod_formats = ['docx']
 itemTemplate.pod_portal_types = ['MeetingItem']
 itemTemplate.tal_condition = u'python:tool.isManager(cfg)'
 
 itemTemplatePDF = PodTemplateDescriptor('deliberation-2', 'Délibération')
-itemTemplatePDF.pod_template_to_use = {'cfg_id': collegeMeeting.id, 'template_id': itemTemplate.id}
+itemTemplatePDF.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': itemTemplate.id}
 itemTemplatePDF.pod_formats = ['pdf']
 itemTemplatePDF.pod_portal_types = ['MeetingItem']
 
 dfAdviceTemplate = PodTemplateDescriptor('avis-df', 'Avis DF')
 dfAdviceTemplate.is_reusable = True
 dfAdviceTemplate.odt_file = '../../examples_fr/templates/avis-df.odt'
-dfAdviceTemplate.pod_formats = ['pdf', ]
+dfAdviceTemplate.pod_formats = ['pdf']
 dfAdviceTemplate.pod_portal_types = ['MeetingItem']
 dfAdviceTemplate.tal_condition = u'python: context.adapted().showFinanceAdviceTemplate()'
 
 dashboardExportTemplate = PodTemplateDescriptor('export', 'Export', dashboard=True)
 dashboardExportTemplate.is_reusable = True
 dashboardExportTemplate.odt_file = '../../examples_fr/templates/dashboard.ods'
-dashboardExportTemplate.pod_formats = ['xlsx', ]
+dashboardExportTemplate.pod_formats = ['xlsx']
 dashboardExportTemplate.tal_condition = u'python: context.absolute_url().endswith("/searches_items")'
 
-dashboardDFTemplateOds = PodTemplateDescriptor('synthese-avis-df-ods', 'Synthèse avis DF', dashboard=True)
+dashboardDFTemplateOds = PodTemplateDescriptor(
+    'synthese-avis-df-ods', 'Synthèse avis DF', dashboard=True)
 dashboardDFTemplateOds.is_reusable = True
 dashboardDFTemplateOds.odt_file = '../../examples_fr/templates/synthese-df-tb.ods'
-dashboardDFTemplateOds.pod_formats = ['xlsx', ]
+dashboardDFTemplateOds.pod_formats = ['xlsx']
 dashboardDFTemplateOds.dashboard_collections_ids = [FINANCE_ADVICES_COLLECTION_ID]
 
 dashboardMeetingAttendances = PodTemplateDescriptor(
@@ -191,9 +197,10 @@ dashboardMeetingAttendances.tal_condition = u'python:False'
 dashboardMeetingAttendances.roles_bypassing_talcondition = set(['Manager', 'MeetingManager'])
 dashboardMeetingAttendances.dashboard_collections_ids = ['searchallmeetings']
 
-collegeTemplates = [dfAdviceTemplate, dashboardExportTemplate, dashboardDFTemplateOds,
-                    dashboardMeetingAttendances, agendaTemplate, agendaTemplatePDF, decisionsTemplate,
-                    decisionsTemplatePDF, itemTemplate, itemTemplatePDF]
+collegeTemplates = [
+    dfAdviceTemplate, dashboardExportTemplate, dashboardDFTemplateOds,
+    dashboardMeetingAttendances, agendaTemplate, agendaTemplatePDF, decisionsTemplate,
+    decisionsTemplatePDF, itemTemplate, itemTemplatePDF]
 
 collegeMeeting.customAdvisers = [
     {'row_id': 'unique_id_002',
@@ -297,7 +304,7 @@ collegeMeeting.recurringItems = [
         title='Approuve le procès-verbal de la séance précédente',
         description='',
         proposingGroup='dirgen',
-        decision='Procès-verbal approuvé'), ]
+        decision='Procès-verbal approuvé')]
 collegeMeeting.itemTemplates = []
 collegeMeeting.initItemDecisionIfEmptyOnDecide = False
 collegeMeeting.meetingPresentItemWhenNoCurrentMeetingStates = ("created", "frozen")
@@ -315,7 +322,8 @@ agendaTemplateCouncil.pod_portal_types = ['Meeting']
 agendaTemplateCouncil.tal_condition = u'python:tool.isManager(cfg)'
 
 agendaTemplatePDFCouncil = PodTemplateDescriptor('oj-2', 'Ordre du jour')
-agendaTemplatePDFCouncil.pod_template_to_use = {'cfg_id': councilMeeting.id, 'template_id': agendaTemplateCouncil.id}
+agendaTemplatePDFCouncil.pod_template_to_use = {
+    'cfg_id': councilMeeting.id, 'template_id': agendaTemplateCouncil.id}
 agendaTemplatePDFCouncil.pod_formats = ['pdf']
 agendaTemplatePDFCouncil.pod_portal_types = ['Meeting']
 
@@ -327,44 +335,48 @@ decisionsTemplateCouncil.pod_portal_types = ['Meeting']
 decisionsTemplateCouncil.tal_condition = u'python:tool.isManager(cfg)'
 
 decisionsTemplatePDFCouncil = PodTemplateDescriptor('pv-2', 'Procès-verbal')
-decisionsTemplatePDFCouncil.pod_template_to_use = {'cfg_id': councilMeeting.id, 'template_id': decisionsTemplateCouncil.id}
+decisionsTemplatePDFCouncil.pod_template_to_use = {
+    'cfg_id': councilMeeting.id, 'template_id': decisionsTemplateCouncil.id}
 decisionsTemplatePDFCouncil.pod_formats = ['pdf']
 decisionsTemplatePDFCouncil.pod_portal_types = ['Meeting']
 
 itemTemplateCouncil = PodTemplateDescriptor('deliberation', 'Délibération')
 itemTemplateCouncil.is_reusable = True
 itemTemplateCouncil.odt_file = '../../examples_fr/templates/deliberation.odt'
-itemTemplateCouncil.pod_formats = ['docx' ]
+itemTemplateCouncil.pod_formats = ['docx']
 itemTemplateCouncil.pod_portal_types = ['MeetingItem']
 itemTemplateCouncil.tal_condition = u'python:tool.isManager(cfg)'
 
 itemTemplatePDFCouncil = PodTemplateDescriptor('deliberation-2', 'Délibération')
-itemTemplatePDFCouncil.pod_template_to_use = {'cfg_id': councilMeeting.id, 'template_id': itemTemplateCouncil.id}
+itemTemplatePDFCouncil.pod_template_to_use = {
+    'cfg_id': councilMeeting.id, 'template_id': itemTemplateCouncil.id}
 itemTemplatePDFCouncil.pod_formats = ['pdf']
 itemTemplatePDFCouncil.pod_portal_types = ['MeetingItem']
 
 dfAdviceTemplateCouncil = PodTemplateDescriptor('avis-df', 'Avis DF')
-dfAdviceTemplateCouncil.pod_template_to_use = {'cfg_id': collegeMeeting.id, 'template_id': dfAdviceTemplate.id}
-dfAdviceTemplateCouncil.pod_formats = ['pdf', ]
+dfAdviceTemplateCouncil.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': dfAdviceTemplate.id}
+dfAdviceTemplateCouncil.pod_formats = ['pdf']
 dfAdviceTemplateCouncil.pod_portal_types = ['MeetingItem']
 dfAdviceTemplateCouncil.tal_condition = u'python: context.adapted().showFinanceAdviceTemplate()'
 
 dashboardExportTemplateCouncil = PodTemplateDescriptor('export', 'Export', dashboard=True)
 dashboardExportTemplateCouncil.pod_template_to_use = {'cfg_id': collegeMeeting.id,
                                                       'template_id': dashboardExportTemplate.id}
-dashboardExportTemplateCouncil.pod_formats = ['xlsx', ]
+dashboardExportTemplateCouncil.pod_formats = ['xlsx']
 dashboardExportTemplateCouncil.tal_condition = u'python: context.absolute_url().endswith("/searches_items")'
 
-dashboardDFTemplateOdsCouncil = PodTemplateDescriptor('synthese-avis-df-ods', 'Synthèse avis DF', dashboard=True)
-dashboardDFTemplateOdsCouncil.pod_template_to_use = {'cfg_id': collegeMeeting.id,
-                                                     'template_id': dashboardDFTemplateOds.id}
-dashboardDFTemplateOdsCouncil.pod_formats = ['xlsx', ]
+dashboardDFTemplateOdsCouncil = PodTemplateDescriptor(
+    'synthese-avis-df-ods', 'Synthèse avis DF', dashboard=True)
+dashboardDFTemplateOdsCouncil.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': dashboardDFTemplateOds.id}
+dashboardDFTemplateOdsCouncil.pod_formats = ['xlsx']
 dashboardDFTemplateOdsCouncil.dashboard_collections_ids = [FINANCE_ADVICES_COLLECTION_ID]
 
 dashboardMeetingAttendancesCouncil = PodTemplateDescriptor(
     'attendance-stats', 'Statistiques de présences', dashboard=True)
-dashboardMeetingAttendancesCouncil.pod_template_to_use = {'cfg_id': collegeMeeting.id,
-                                                     'template_id': dashboardMeetingAttendances.id}
+dashboardMeetingAttendancesCouncil.pod_template_to_use = {
+    'cfg_id': collegeMeeting.id, 'template_id': dashboardMeetingAttendances.id}
 dashboardMeetingAttendancesCouncil.pod_formats = ['xlsx']
 dashboardMeetingAttendancesCouncil.tal_condition = u'python:False'
 dashboardMeetingAttendancesCouncil.roles_bypassing_talcondition = set(['Manager', 'MeetingManager'])
@@ -372,14 +384,15 @@ dashboardMeetingAttendancesCouncil.dashboard_collections_ids = ['searchallmeetin
 
 pubTemplate = PodTemplateDescriptor('publications', 'Publications (www.deliberations.be)')
 pubTemplate.odt_file = '../../examples_fr/templates/publications.odt'
-pubTemplate.pod_formats = ['pdf' ]
+pubTemplate.pod_formats = ['pdf']
 pubTemplate.pod_portal_types = ['Meeting']
 pubTemplate.tal_condition = u'python:tool.isManager(cfg)'
 
-councilTemplates = [dfAdviceTemplateCouncil, dashboardExportTemplateCouncil,
-                    dashboardDFTemplateOdsCouncil, dashboardMeetingAttendancesCouncil,
-                    agendaTemplateCouncil, agendaTemplatePDFCouncil, decisionsTemplateCouncil,
-                    decisionsTemplatePDFCouncil, itemTemplateCouncil, itemTemplatePDFCouncil, pubTemplate]
+councilTemplates = [
+    dfAdviceTemplateCouncil, dashboardExportTemplateCouncil,
+    dashboardDFTemplateOdsCouncil, dashboardMeetingAttendancesCouncil,
+    agendaTemplateCouncil, agendaTemplatePDFCouncil, decisionsTemplateCouncil,
+    decisionsTemplatePDFCouncil, itemTemplateCouncil, itemTemplatePDFCouncil, pubTemplate]
 
 councilMeeting.shortName = 'Council'
 councilMeeting.assembly = ''
@@ -409,9 +422,10 @@ councilMeeting.insertingMethodsOnAddItem = (
     {'insertingMethod': 'on_proposing_groups', 'reverse': '0'})
 councilMeeting.xhtmlTransformFields = ()
 councilMeeting.xhtmlTransformTypes = ()
-councilMeeting.itemReferenceFormat = "python: 'CC/' + (here.hasMeeting() and " \
-                                     "here.restrictedTraverse('@@pm_unrestricted_methods').getLinkedMeetingDate().strftime('%Y%m%d') or '') " \
-                                     "+ '-' + str(here.getItemNumber(relativeTo='meeting', for_display=True))"
+councilMeeting.itemReferenceFormat = \
+    "python: 'CC/' + (here.hasMeeting() and " \
+    "here.restrictedTraverse('@@pm_unrestricted_methods').getLinkedMeetingDate().strftime('%Y%m%d') or '') " \
+    "+ '-' + str(here.getItemNumber(relativeTo='meeting', for_display=True))"
 councilMeeting.itemWFValidationLevels = (
     {'state': 'itemcreated',
      'state_title': 'itemcreated',
@@ -467,8 +481,8 @@ councilMeeting.itemsListVisibleColumns = [
     u'static_item_reference', u'Creator', u'CreationDate', u'review_state', u'getCategory',
     u'getProposingGroup', u'advices', u'actions']
 councilMeeting.enabledAnnexesBatchActions = ['delete', 'download-annexes']
-councilMeeting.dashboardItemsListingsFilters = ('c4', 'c5', 'c6', 'c7', 'c8', 'c9', 'c10',
-                                              'c11', 'c13', 'c14', 'c15', 'c29', 'c32')
+councilMeeting.dashboardItemsListingsFilters = (
+    'c4', 'c5', 'c6', 'c7', 'c8', 'c9', 'c10', 'c11', 'c13', 'c14', 'c15', 'c29', 'c32')
 councilMeeting.dashboardMeetingAvailableItemsFilters = ('c4', 'c5', 'c11', 'c29', 'c32')
 councilMeeting.dashboardMeetingLinkedItemsFilters = ('c4', 'c5', 'c6', 'c7', 'c11', 'c19', 'c29', 'c32')
 councilMeeting.useAdvices = True
@@ -485,8 +499,9 @@ councilMeeting.itemAdviceViewStates = ('proposed',
                                        'accepted_but_modified',
                                        'refused',
                                        'delayed')
-councilMeeting.usedAdviceTypes = ['positive', 'positive_with_comments', 'positive_with_remarks', 'cautious',
-                                  'negative', 'negative_with_remarks', 'back_to_proposing_group', 'nil', 'read']
+councilMeeting.usedAdviceTypes = [
+    'positive', 'positive_with_comments', 'positive_with_remarks', 'cautious',
+    'negative', 'negative_with_remarks', 'back_to_proposing_group', 'nil', 'read']
 collegeMeeting.transitionsReinitializingDelays = ['backToItemCreated']
 councilMeeting.keepAccessToItemWhenAdviceIsGiven = True
 councilMeeting.inheritedAdviceRemoveableByAdviser = True
@@ -552,7 +567,7 @@ councilMeeting.recurringItems = [
         description='',
         category='administration',
         proposingGroup='dirgen',
-        decision='Procès-verbal approuvé'), ]
+        decision='Procès-verbal approuvé')]
 councilMeeting.itemTemplates = []
 councilMeeting.itemIconColor = "orange"
 councilMeeting.initItemDecisionIfEmptyOnDecide = False
@@ -565,3 +580,4 @@ data = PloneMeetingConfiguration(
     orgs=groups)
 data.usersOutsideGroups = []
 data.directory_position_types = list(examples_fr_import_data.data.directory_position_types)
+data.contactsTemplates = list(examples_fr_import_data.data.contactsTemplates)